moved useDashboard to hooks dir
